feat: use google sheets to manage persistent data
The admin would make a google sheet and share the link of the sheet in
editor mode. Then, using that link, on the Admin page, the admin would
create a QR code for users to scan. A user can scan the qr code and
they'll be brought to the App page where they can do their stuff.

fixes #5